Send Mails once per Status and extract MailConfiguration in own class (#3)

+++ ServerMonitoring.Application/Responses/TemperatureResponse.cs
@@ -2,7 +2,7 @@ namespace ServerMonitoring.Application.Responses;
 
 public class TemperatureResponse
 {
-    public int Temperature { get; init; }
+    public double Temperature { get; init; }
     public TemperatureStatus Status { get; init; }
 }
